+++ R/summaries.R
@@ -1,6 +1,6 @@
 #' Create numerical summaries
 #'
-#' `num_sum()` is a wrapper around `mosaic::favstats()` and creates a tidy table of summary statistics
+#' `tbl_num_sum()` is a wrapper around `mosaic::favstats()` and creates a tidy table of summary statistics
 #'    including the min, q1, median, q3, max, mean, standard deviation, n, and missing values.
 #'
 #' @param data A data frame (or tibble).
@@ -16,12 +16,12 @@
 #'
 #' @examples
 #' # not removing NAs is not recommended
-#' num_sum(mtcars, ~wt)
+#' tbl_num_sum(mtcars, ~wt)
 #'
-#' num_sum(mtcars, ~wt, na_rm = TRUE)
-#' num_sum(mtcars, ~wt, na_rm = TRUE, digits = 2, caption = "This is a table")
-#' num_sum(mtcars, wt~cyl, na_rm = TRUE)
-num_sum <- function(data, formula, digits = 3, na_rm = FALSE, caption = NULL) {
+#' tbl_num_sum(mtcars, ~wt, na_rm = TRUE)
+#' tbl_num_sum(mtcars, ~wt, na_rm = TRUE, digits = 2, caption = "This is a table")
+#' tbl_num_sum(mtcars, wt~cyl, na_rm = TRUE)
+tbl_num_sum <- function(data, formula, digits = 3, na_rm = FALSE, caption = NULL) {
 
   # error catching
   if (na_rm == FALSE) {
@@ -75,10 +75,10 @@ num_sum <- function(data, formula, digits = 3, na_rm = FALSE, caption = NULL) {
 
 #' Create percentile summaries
 #'
-#' `pctile()` is a wrapper around `mosaic::quantile()` and creates a tidy table of data values
+#' `tbl_pctile()` is a wrapper around `mosaic::quantile()` and creates a tidy table of data values
 #'    at the given percentiles.
 #'
-#' @inheritParams num_sum
+#' @inheritParams tbl_num_sum
 #' @param probs A vector of percentiles to compute. Each value must be between 0 and 1, inclusive.
 #'    Defaults to 0, 25%, 50%, 75%, and 100%.
 #'
@@ -87,12 +87,12 @@ num_sum <- function(data, formula, digits = 3, na_rm = FALSE, caption = NULL) {
 #' @export
 #'
 #' @examples
-#' pctile(mtcars, ~wt)
-#' pctile(mtcars, ~wt, probs = c(.17, .3, .5, .7, .9, 1), na_rm = TRUE)
-#' pctile(mtcars, wt~cyl, na_rm = TRUE)
+#' tbl_pctile(mtcars, ~wt)
+#' tbl_pctile(mtcars, ~wt, probs = c(.17, .3, .5, .7, .9, 1), na_rm = TRUE)
+#' tbl_pctile(mtcars, wt~cyl, na_rm = TRUE)
 #'
-#' try(pctile(mtcars, ~wt, probs = c(25, 50, 75, 100)))
-pctile <- function(data, formula, digits = 3, probs = c(0, .25, .5, .75, 1), caption = NULL, na_rm = FALSE) {
+#' try(tbl_pctile(mtcars, ~wt, probs = c(25, 50, 75, 100)))
+tbl_pctile <- function(data, formula, digits = 3, probs = c(0, .25, .5, .75, 1), caption = NULL, na_rm = FALSE) {
 
   # error catching
 
@@ -151,7 +151,7 @@ pctile <- function(data, formula, digits = 3, probs = c(0, .25, .5, .75, 1), cap
 
 #' Create a simple correlation table
 #'
-#' @inheritParams num_sum
+#' @inheritParams tbl_num_sum
 #' @param formula Two variables given in formula notation: `var1~var2`.
 #'
 #' @return An object of class flextable. If in an interactive session, the table will be
@@ -160,10 +160,10 @@ pctile <- function(data, formula, digits = 3, probs = c(0, .25, .5, .75, 1), cap
 #'
 #' @examples
 #' # Not removing NAs is not recommended
-#' corr(mtcars, wt~qsec)
+#' tbl_corr(mtcars, wt~qsec)
 #'
-#' corr(mtcars, wt~qsec, na_rm = TRUE)
-corr <- function(data, formula, digits = 3, caption = NULL, na_rm = FALSE) {
+#' tbl_corr(mtcars, wt~qsec, na_rm = TRUE)
+tbl_corr <- function(data, formula, digits = 3, caption = NULL, na_rm = FALSE) {
 
   # error catching
   if (na_rm == FALSE) {